refactor: pull types out of amaru-plutus, leaving just the serialization logic
