+++ crates/amaru-plutus/src/script_context/v3.rs
@@ -16,27 +16,28 @@ use std::{collections::BTreeMap, ops::Deref};
 
 use amaru_kernel::{
     Address, AssetName, Bytes, Certificate as PallasCertificate, ComparableProposalId, Constitution, CostModels, DRep,
-    DRepVotingThresholds, ExUnitPrices, ExUnits, GovernanceAction, Hash, PoolVotingThresholds, Proposal, ProposalId,
-    ProtocolParamUpdate, RationalNumber, StakeCredential, StakePayload, Vote, size::CREDENTIAL,
+    DRepVotingThresholds, ExUnitPrices, ExUnits, GovernanceAction, Hash, PlutusData, PoolVotingThresholds, Proposal,
+    ProposalId, ProtocolParamUpdate, RationalNumber, StakeCredential, StakePayload, TransactionInput, Vote, Voter,
+    size::CREDENTIAL,
 };
 use num::Integer;
 
 use crate::{
     PlutusDataError, ToPlutusData, constr, constr_v3,
     script_context::{
-        Certificate, CurrencySymbol, Datums, Mint, OutputRef, PlutusData, ScriptContext, ScriptInfo, ScriptPurpose,
-        StakeAddress, TransactionInput, TransactionOutput, TxInfo, Value, Voter, Votes, Withdrawals,
+        Certificate, CurrencySymbol, Datums, Mint, OutputReference, ScriptContext, ScriptInfo, ScriptPurpose,
+        StakeAddress, TransactionOutput, TxInfo, Value, Votes, Withdrawals,
     },
 };
 
-impl ToPlutusData<3> for OutputRef<'_> {
-    /// Serialize an `OutputRef` as PlutusData for PlutusV3.
+impl ToPlutusData<3> for OutputReference<'_> {
+    /// Serialize an `OutputReference` as PlutusData for PlutusV3.
     ///
     /// # Errors
     /// If the UTxO is locked at a bootstrap address, this will return a `PlutusDataError`.
     fn to_plutus_data(&self) -> Result<PlutusData, PlutusDataError> {
         if let Address::Byron(_) = *self.output.address {
-            return Err(PlutusDataError::unsupported_version("byron address included in OutputRef", 3));
+            return Err(PlutusDataError::unsupported_version("byron address included in OutputReference", 3));
         }
 
         constr_v3!(0, [self.input, self.output])
